+++ src/common.h
@@ -4,21 +4,21 @@
 
 // Constants
 
-constexpr double infinity = std::numeric_limits<double>::infinity();
-constexpr double pi = 3.1415926535897932385;
+constexpr float infinity = std::numeric_limits<float>::infinity();
+constexpr float pi = 3.1415926535897932385;
 
 // Utility Functions
 
-inline constexpr double degrees_to_radians(double degrees) {
+inline constexpr float degrees_to_radians(float degrees) {
     return degrees * pi / 180.0;
 }
 
-inline double random_double() {
-    static std::uniform_real_distribution<double> distribution(0.0, 1.0);
+inline float random_double() {
+    static std::uniform_real_distribution<float> distribution(0.0, 1.0);
     static std::mt19937 generator;
     return distribution(generator);
 }
 
-inline double random_double(double min, double max) {
+inline float random_double(float min, float max) {
     return min + (max - min) * random_double();
 }
