+++ src/ray.h
@@ -13,7 +13,7 @@ public:
     const point3& origin() const { return orig; }
     const vec3& direction() const { return dir; }
 
-    vec3 at(double t) const {
+    vec3 at(float t) const {
         return N_SSE2::point_at(orig.vec, dir.vec, t);
     }
